Add files via upload
This is the Task 1 of the Internship Program where we have to implement  Advanced Data Structures, a novel version of a heap.
Your task is to implement a novel data structure - your project lead is calling it a power of two max heap. The rest of your team is doing their best to come up with a better name. The requirements of the data structure are as follows:

The heap must satisfy the heap property.
Every parent node in the heap must have 2^x children.
The value of x must be a parameter of the heap’s constructor.
The heap must implement an insert method.
The heap must implement a pop max method.
The heap must be implemented in Java.
The heap must be performant.
You must use a more descriptive variable name than x in your implementation.